[backend] refactor: rename user_service files and update import paths to user-service
