+++ apps/portal-api/src/modules/security-management/service-capability/service-capability.app.ts
@@ -4,7 +4,7 @@ import { ServiceInstanceId } from '../../../model/kanel/public/ServiceInstance';
 import { UserServiceId } from '../../../model/kanel/public/UserService';
 import { assertUserCanManageService } from '../../../security/guard';
 import { fillSubscriptionWithOrgaServiceAndUserService } from '../../subscription/subscription.domain';
-import { UserServiceDomain } from '../../user_service/user_service.domain';
+import { UserServiceDomain } from '../../user-service/user-service.domain';
 import { insertCapabilities } from '../user-service-capability/user-service-capability.helper';
 
 import { willManageAccessBeConserved } from './service-capability.helper';
