+++ apps/portal-api/src/modules/security-management/service-capability/service-capability.helper.ts
@@ -4,7 +4,7 @@ import { requestContext } from '../../../context/request.context';
 import { UserServiceId } from '../../../model/kanel/public/UserService';
 import UserServiceCapability from '../../../model/kanel/public/UserServiceCapability';
 import { ErrorCode } from '../../../utils/error/error.code';
-import { UserServiceDomain } from '../../user_service/user_service.domain';
+import { UserServiceDomain } from '../../user-service/user-service.domain';
 import { GenericServiceCapabilityName } from './generic-service-capability.const';
 import { getManageAccessLeft } from './service-capability.domain';
