+++ apps/portal-api/src/modules/subscription/subscription.domain.ts
@@ -14,7 +14,7 @@ import { restrictSubscriptionToUserOrganization } from '../../security/restricti
 import { loadOrganizationBy } from '../organization-management/organizations/organizations.domain';
 import { loadUserBy } from '../organization-management/users/user-domain/users.domain';
 import { loadServiceInstanceBy } from '../service/instance/service-instance.domain';
-import { UserServiceDomain } from '../user_service/user_service.domain';
+import { UserServiceDomain } from '../user-service/user-service.domain';
 import { loadSubscriptionWithOrganizationAndCapabilitiesBy } from './subscription.helper';
 
 export const SubscriptionDomain = {
