+++ apps/portal-api/src/modules/user-service/user-service.domain.ts
