+++ apps/portal-api/src/modules/user-service/user-service.resolver.ts
@@ -6,8 +6,8 @@ import { UnknownErrorCode } from '../../utils/error/error.code';
 import { mapToGraphQLError } from '../../utils/error/error.mapping';
 import { extractId } from '../../utils/utils';
 import { loadUserDetails } from '../organization-management/users/user-domain/users.domain';
-import { UserServiceApp } from './user_service.app';
-import { UserServiceDomain } from './user_service.domain';
+import { UserServiceApp } from './user-service.app';
+import { UserServiceDomain } from './user-service.domain';
 
 const resolvers: Resolvers = {
   UserService: {
